refactor commands (delete old)
