+++ modules/command/common/deposit-tokens-command.js
@@ -1,4 +1,4 @@
-const BN = require('bn.js');
+const BN = require('../../../node_modules/bn.js/lib/bn');
 
 const Command = require('../command');
 const Utilities = require('../../Utilities');
