+++ modules/command/common/pay-out-command.js
