+++ modules/command/dc/dc-offer-mining-completed-command.js
@@ -1,4 +1,4 @@
-const BN = require('bn.js');
+const BN = require('../../../node_modules/bn.js/lib/bn');
 const Command = require('../command');
 const models = require('../../../models/index');
