+++ modules/command/dc/dc-offer-mining-status-command.js
