+++ modules/command/dc/dc-replication-request-command.js
@@ -1,4 +1,4 @@
-const BN = require('bn.js');
+const BN = require('../../../node_modules/bn.js/lib/bn');
 const path = require('path');
 
 const Command = require('../command');
